+++ app/src/main/java/com/example/appdatcomtam/Screen/ScreenUser/HomeUser.kt
@@ -1,0 +1,10 @@
+package com.example.appdatcomtam.Screen.ScreenUser
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.navigation.NavController
+
+@Composable
+fun HomeUser(navController: NavController? = null,){
+Text(text = "Đây là màn hình User")
+}
